+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Src
 - 📈 **markdown viewer** now renders YAML headers in a box, but now disabled by default - to show the header you supply the `--header` flag
+- 📈 now **.url** files are also supported for image preview if it contains a `.ico` IconFile
 
 ## V0.4.5
 - 📈 **markdown viewer** now supports `<figure>` and `<figcaption>` HTML elements